+++ app/src/main/java/app/kaito_dogi/mybrary/MainActivity.kt
@@ -12,7 +12,7 @@ import app.kaito_dogi.mybrary.core.designsystem.theme.MybraryTheme
 import app.kaito_dogi.mybrary.core.ui.exception.ExceptionConsumer
 import app.kaito_dogi.mybrary.core.ui.exception.ExceptionConsumerEntryPoint
 import app.kaito_dogi.mybrary.core.ui.navigation.AppNavHost
-import app.kaito_dogi.mybrary.core.ui.navigation.mainNavGraph
+import app.kaito_dogi.mybrary.core.ui.navigation.bar.mainNavGraph
 import app.kaito_dogi.mybrary.core.ui.navigation.route.AppRoute
 import app.kaito_dogi.mybrary.core.ui.navigation.route.MainRoute
 import app.kaito_dogi.mybrary.core.ui.navigation.route.MyBookRoute
