+++ core/ui/src/main/java/app/kaito_dogi/mybrary/core/ui/navigation/AppNavHost.kt
@@ -16,6 +16,8 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
+import app.kaito_dogi.mybrary.core.ui.navigation.bar.MainNavigationBar
+import app.kaito_dogi.mybrary.core.ui.navigation.bar.NavigationBarDestination
 import app.kaito_dogi.mybrary.core.ui.navigation.route.AppRoute
 import app.kaito_dogi.mybrary.core.ui.navigation.route.MainRoute
