+++ core/ui/src/main/java/app/kaito_dogi/mybrary/core/ui/navigation/bar/NavigationBarDestination.kt
@@ -1,4 +1,4 @@
-package app.kaito_dogi.mybrary.core.ui.navigation
+package app.kaito_dogi.mybrary.core.ui.navigation.bar
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
